Add component attribute to Dependency model and class
The code changes add a new `component` attribute to the `Dependency` model and class in both the `models.py` and `dependency.py` files. This attribute is used to store the SBOM linked attributes of the dependency. The `component` attribute is included in the JSON representation of the `Dependency` model and class.